chore: Renames information elements to ie
