+++ client/client_test.go
@@ -5,7 +5,7 @@ import (
 	"time"
 
 	"github.com/dot-5g/pfcp/client"
-	"github.com/dot-5g/pfcp/information_elements"
+	"github.com/dot-5g/pfcp/ie"
 )
 
 type MockUdpSender struct {
@@ -24,7 +24,7 @@ func TestGivenPfcpWhenSendHeartbeatRequestThenNoError(t *testing.T) {
 	}
 	pfcpClient := client.New("127.0.0.1:8805")
 	pfcpClient.Udp = mockSender
-	recoveryTimeStamp := information_elements.NewRecoveryTimeStamp(time.Now())
+	recoveryTimeStamp := ie.NewRecoveryTimeStamp(time.Now())
 	sequenceNumber := uint32(21)
 
 	_, err := pfcpClient.SendHeartbeatRequest(recoveryTimeStamp, sequenceNumber)
