--- ie/nodeID.go
@@ -1,10 +1,0 @@
-package information_elements
-
-type NodeIDType int
-
-type NodeID struct {
-	Type        int
-	Length      int
-	NodeIDType  NodeIDType
-	NodeIDValue []byte
-}
